:truck: Separate SHMConfig and SREConfig into their own files
