+++ data_safe_haven/config/__init__.py
@@ -1,6 +1,7 @@
-from .config import SHMConfig, SREConfig
 from .dsh_pulumi_config import DSHPulumiConfig
 from .dsh_pulumi_project import DSHPulumiProject
+from .shm_config import SHMConfig
+from .sre_config import SREConfig
 
 __all__ = [
     "DSHPulumiConfig",
