+++ data_safe_haven/config/config_sections.py
@@ -1,16 +1,12 @@
-"""Configuration file backed by blob storage"""
+"""Sections for use in configuration files"""
 
 from __future__ import annotations
-
-from typing import ClassVar, Self
 
 from pydantic import (
     BaseModel,
     Field,
 )
 
-from data_safe_haven.functions import sanitise_sre_name
-from data_safe_haven.serialisers import AzureSerialisableModel, ContextBase
 from data_safe_haven.types import (
     AzureVmSku,
     DatabaseSystem,
@@ -180,75 +176,3 @@ class ConfigSectionSRE(BaseModel, validate_assignment=True):
         if workspace_skus:
             self.workspace_skus = workspace_skus
         logger.info(f"[bold]Workspace SKUs[/] will be [green]{self.workspace_skus}[/].")
-
-
-class SHMConfig(AzureSerialisableModel):
-    config_type: ClassVar[str] = "SHMConfig"
-    filename: ClassVar[str] = "shm.yaml"
-    azure: ConfigSectionAzure
-    shm: ConfigSectionSHM
-
-    def is_complete(self) -> bool:
-        if not all((self.azure, self.shm)):
-            return False
-        return True
-
-    @classmethod
-    def template(cls: type[Self]) -> SHMConfig:
-        """Create object without validation to allow "replace me" prompts."""
-        return SHMConfig.model_construct(
-            azure=ConfigSectionAzure.model_construct(
-                subscription_id="Azure subscription ID",
-                tenant_id="Azure tenant ID",
-            ),
-            shm=ConfigSectionSHM.model_construct(
-                entra_tenant_id="Entra tenant ID",
-                fqdn="TRE domain name",
-            ),
-        )
-
-
-class SREConfig(AzureSerialisableModel):
-    config_type: ClassVar[str] = "Config"
-    filename: ClassVar[str] = "config.yaml"
-    azure: ConfigSectionAzure
-    sre: ConfigSectionSRE
-
-    def is_complete(self) -> bool:
-        if not all((self.azure, self.sre)):
-            return False
-        return True
-
-    @classmethod
-    def sre_from_remote(cls: type[Self], context: ContextBase, sre_name: str) -> SREConfig:
-        """Load a Config from Azure storage."""
-        return cls.from_remote(context, filename=cls.sre_filename_from_name(sre_name))
-
-    @classmethod
-    def sre_filename_from_name(cls: type[Self], sre_name: str) -> str:
-        """Construct a canonical filename."""
-        return f"sre-{sanitise_sre_name(sre_name)}.yaml"
-
-    @classmethod
-    def template(cls: type[Self]) -> SREConfig:
-        """Create object without validation to allow "replace me" prompts."""
-        return SREConfig.model_construct(
-            azure=ConfigSectionAzure.model_construct(
-                subscription_id="Azure subscription ID",
-                tenant_id="Azure tenant ID",
-            ),
-            sre=ConfigSectionSRE.model_construct(
-                admin_email_address="Admin email address",
-                databases=["List of database systems to enable"],
-                data_provider_ip_addresses=["Data provider IP addresses"],
-                remote_desktop=ConfigSubsectionRemoteDesktopOpts.model_construct(
-                    allow_copy="Whether to allow copying text out of the environment",
-                    allow_paste="Whether to allow pasting text into the environment",
-                ),
-                workspace_skus=[
-                    "Azure VM SKUs - see cloudprice.net for list of valid SKUs"
-                ],
-                research_user_ip_addresses=["Research user IP addresses"],
-                software_packages=SoftwarePackageCategory.ANY,
-            ),
-        )
